pyproject.toml
[project]
name = "paper-summarizer"
version = "0.1.0"
description = "A web application for summarizing research papers"
authors = [{name = "gr8monk3ys"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0",
    "alembic>=1.13.0",
    "arq>=0.26.0",
    "prometheus-client>=0.20.0",
    "sentry-sdk>=2.8.0",
    "sqlmodel>=0.0.21",
    "uvicorn[standard]>=0.30.0",
    "reportlab>=4.2.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "beautifulsoup4>=4.9.3",
    "requests>=2.25.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "together>=0.2.0",
]

[project.optional-dependencies]
local = [
    "nltk>=3.8.0",
    "sumy>=0.11.0",
    "gensim>=4.3.3",
    "transformers>=4.30.0",
    "torch>=2.0.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
